// src/app/shared/interfaces/datos-historia.ts
export interface DatosHistoria {
  documento: string;
  tipo_documento: string;
  nombres: string;
  apellidos: string;
  fecha_nacimiento: string;
  lugar_nacimiento: string;
  nacionalidad: string;
  genero: string;
  direccion: string;
  telefono: string;
  empresa: string;
  cargo: string;
  fecha_ingreso: string;
  tiempo_cargo: string;
  arl: string;
  eps: string;
  afp: string;
  telefono_empresa: string;
  correo: string;
  password: string;
  foto: string;
}